add external exclude files and updates with design

+++ Makefile
@@ -4,22 +4,25 @@
 #
 
 # Colors
-NO_COLOR=\033[0m
-TARGET_COLOR=\033[32;01m
-OK_COLOR=\033[32;01m
-ERROR_COLOR=\033[31;01m
-WARN_COLOR=\033[33;01m
-ACTION=$(TARGET_COLOR)--> 
+NO_COLOR		= \033[0m
+TARGET_COLOR	= \033[32;01m
+OK_COLOR		= \033[32;01m
+ERROR_COLOR		= \033[31;01m
+WARN_COLOR		= \033[33;01m
+ACTION			= $(TARGET_COLOR)--> 
 
 # Add local bin path for test tools
-PATH := "$(PWD)/bin:$(PWD)/vendor/bin:$(PWD)/node_modules/.bin:$(PATH)"
+BIN 		= bin
+VENDORBIN 	= vendor/bin
+NPMBIN		= node_modules/.bin
+PATH		:= $(BIN):$(VENDORBIN):$(NPMBIN):$(PATH)
 
 
 
 # target: help          - Displays help.
 .PHONY:  help
 help:
-	@echo "Displaying help for this Makefile."
+	@echo "$(ACTION)Displaying help for this Makefile.$(NO_COLOR)"
 	@echo "Usage:"
 	@echo " make [target] ..."
 	@echo "target:"
@@ -67,9 +70,9 @@ clean-all: clean
 .PHONY: dbwebb-install
 dbwebb-install: build-prepare
 	@echo "$(ACTION)Download and install dbwebb$(NO_COLOR)"
-	wget --quiet -O bin/dbwebb https://raw.githubusercontent.com/mosbth/dbwebb-cli/master/dbwebb2
-	chmod 755 bin/dbwebb
-	export PATH=$(PATH) && dbwebb config create noinput && dbwebb --version
+	wget --quiet -O $(BIN)/dbwebb https://raw.githubusercontent.com/mosbth/dbwebb-cli/master/dbwebb2
+	chmod 755 $(BIN)/dbwebb
+	$(BIN)/dbwebb config create noinput && $(BIN)/dbwebb --version
 
 
 
@@ -77,7 +80,7 @@ dbwebb-install: build-prepare
 .PHONY: dbwebb-testrepo
 dbwebb-testrepo: dbwebb-install
 	@echo "$(ACTION)Test course repo$(NO_COLOR)"
-	export PATH=$(PATH) && dbwebb --silent --local testrepo
+	PATH=$(PATH); $(BIN)/dbwebb --silent --local testrepo
 
 
 
@@ -85,9 +88,9 @@ dbwebb-testrepo: dbwebb-install
 .PHONY: dbwebb-validate-install
 dbwebb-validate-install: build-prepare
 	@echo "$(ACTION)Download and install dbwebb-validate$(NO_COLOR)"
-	wget --quiet -O bin/dbwebb-validate https://raw.githubusercontent.com/mosbth/dbwebb-cli/master/dbwebb2-validate
-	chmod 755 bin/dbwebb-validate
-	export PATH=$(PATH) && dbwebb-validate --version
+	wget --quiet -O $(BIN)/dbwebb-validate https://raw.githubusercontent.com/mosbth/dbwebb-cli/master/dbwebb2-validate
+	chmod 755 $(BIN)/dbwebb-validate
+	$(BIN)/dbwebb-validate --version
 
 
 
@@ -95,7 +98,7 @@ dbwebb-validate-install: build-prepare
 .PHONY: dbwebb-validate-check
 dbwebb-validate-check:
 	@echo "$(ACTION)Check version and environment for dbwebb-validate$(NO_COLOR)"
-	export PATH=$(PATH) && dbwebb-validate --version && dbwebb-validate --check
+	$(BIN)/dbwebb-validate --version && $(BIN)/dbwebb-validate --check
 
 
 
@@ -103,7 +106,7 @@ dbwebb-validate-check:
 .PHONY: dbwebb-validate-run
 dbwebb-validate-run:
 	@echo "$(ACTION)Run tests with dbwebb-validate$(NO_COLOR)"
-	export PATH=$(PATH) && dbwebb-validate --publish --publish-to build/webroot/ example
+	$(BIN)/dbwebb-validate --publish --publish-to build/webroot/ example
 
 
 
@@ -111,7 +114,7 @@ dbwebb-validate-run:
 .PHONY: npm-install-dev
 npm-install-dev: build-prepare
 	@echo "$(ACTION)Install npm packages for development$(NO_COLOR)"
-	npm install --only=dev
+	if [ -f package.json ]; then npm install --only=dev; fi
 
 
 
@@ -119,7 +122,7 @@ npm-install-dev: build-prepare
 .PHONY: npm-update-dev
 npm-update-dev:
 	@echo "$(ACTION)Update npm packages for development$(NO_COLOR)"
-	npm update --only=dev
+	if [ -f package.json ]; then npm update --only=dev; fi
 
 
 
@@ -185,4 +188,4 @@ test: automated-tests-prepare automated-tests-check automated-tests-run dbwebb-t
 .PHONY: dbwebb-validate
 dbwebb-validate:
 	@echo "$(ACTION)Executed all automated tests$(NO_COLOR)"
-	export PATH=$(PATH) && dbwebb-validate --publish --publish-to build/webroot/ $(arg1)
+	$(BIN)/dbwebb-validate --publish --publish-to build/webroot/ $(arg1)
